limited access to controllers

+++ app/Http/Controllers/API/UserController.php
@@ -142,12 +142,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-
-        // $this->authorize('isAdmin');
+        $this->authorize('isAdmin');
 
         $user = User::findOrFail($id);
         // delete the user
-
         $user->delete();
 
         return ['message' => 'User Deleted'];
